requirements.txt
requests>=2.32.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
openai>=1.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytz>=2024.1